Pairing working again
